requirements.txt
Brotli==1.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
discum==1.4.1
filetype==1.2.0
idna==3.10
input-output-build==0.1.1
python-dotenv==1.2.1
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.4.0
websocket-client==0.59.0
